refactor: change names and add docs for evmSigner for node.js
