+++ warp-contracts-plugin-signature/src/server/index.ts
@@ -1,1 +1,2 @@
 export { EvmSignatureVerificationServerPlugin } from './evm/evmSignatureVerification';
+export { buildEvmSignature } from './evm/buildEvmSignature';
